requirements.txt
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
openai==1.57.0
streamlit==1.40.2
chromadb==0.3.29
tiktoken==0.8.0
python-dotenv==1.0.1
PyMuPDF==1.25.0
docx2txt==0.8
python-docx==1.1.2
beautifulsoup4==4.13.3
tqdm==4.67.1
pandas==2.2.3
numpy==1.26.4